requirements.txt
streamlit==1.31.0
openai==1.12.0
python-dotenv==1.0.0
google-api-python-client==2.116.0
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.2.0
pandas==2.2.0
plotly==5.18.0
streamlit-chat==0.1.1
langchain==0.1.5
langchain-openai==0.0.5
chromadb==0.4.22
tiktoken==0.5.2